pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "shandu"
version = "3.0.0"
description = "Multi-agent research system powered by Blackgeorge"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
  {name = "Dusan Jolovic", email = "jolovic@pm.me"}
]
keywords = ["research", "ai", "blackgeorge", "deepresearch", "search"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
  "blackgeorge==1.1.8",
  "click>=8.1.7",
  "rich>=14.0.0",
  "aiohttp>=3.10.0",
  "beautifulsoup4>=4.12.0",
  "ddgs>=9.0.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
]

[project.scripts]
shandu = "shandu.cli:cli"

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.8.0",
  "mypy>=1.11.0"
]

[tool.hatch.build.targets.wheel]
packages = ["shandu"]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.8.0",
  "mypy>=1.11.0"
]
